add useConfig for emails
