+++ app/src/components/ErrorPage/ErrorPage.tsx
@@ -1,15 +1,17 @@
 import Typography from "@mui/material/Typography";
 import { Link, Box, Button, Stack } from "@mui/material";
 import { useNavigate } from "react-router-dom";
-import { ErrorPageComponent } from "./style";
-import { ADMINISTRATOR_EMAIL, SUPPORT_EMAIL, errors } from "../../hooks/config";
+import { ErrorPageWrapper } from "./style";
+import { errors } from "./errorData";
+import { useConfig } from "../../hooks/config";
 
 export function ErrorPage({ code }: { code: string }): JSX.Element {
   const navigate = useNavigate();
   const errorsMap = new Map(errors.map((e) => [e.code, { ...e }]));
+  const { emails } = useConfig();
 
   return (
-    <ErrorPageComponent>
+    <ErrorPageWrapper>
       <Box className="wrapper">
         <Box className="error_code">
           <Typography
@@ -34,24 +36,25 @@ export function ErrorPage({ code }: { code: string }): JSX.Element {
             </Typography>
             <Typography variant="body2" component="p">
               For technical support, please contact your local administrator
-              <Link href={`mailto:${ADMINISTRATOR_EMAIL}`} variant="body2" className="support_mail">
-                {ADMINISTRATOR_EMAIL}
+              <Link href={`mailto:${emails.admin}`} variant="body2" className="support_mail">
+                {emails.admin}
               </Link>
             </Typography>
             <Typography variant="body2" component="p">
               In case your administrator is unavailable, please contact us for technical support
-              <Link href={`mailto:${SUPPORT_EMAIL}`} variant="body2" className="support_mail">
-                {SUPPORT_EMAIL}
+              <Link href={`mailto:${emails.support}`} variant="body2" className="support_mail">
+                {emails.support}
               </Link>
             </Typography>
           </Stack>
-          {500 <= Number(code) && (
+          {/* Add button `Back to main page' if error's code 500 and more*/}
+          {Number(code) >= 500 && (
             <Button onClick={() => navigate("/")} variant={"outlined"}>
               Back to main page
             </Button>
           )}
         </Box>
       </Box>
-    </ErrorPageComponent>
+    </ErrorPageWrapper>
   );
 }
